+++ Sources/Base/OAuth2Response.swift
@@ -91,7 +91,7 @@ open class OAuth2Response {
 		else if 403 == response.statusCode {
 			throw OAuth2Error.forbidden
 		}
-		else if let data = data {
+		else if let data = data, data.count > 0 {
 			return data
 		}
 		else {