+++ Sources/DataLoader/OAuth2DataLoader.swift
@@ -40,11 +40,17 @@ open class OAuth2DataLoader: OAuth2Requestable {
 	/**
 	Designated initializer.
 	
+	Provide `host` if 301 and 302 redirects should be followed automatically, as long as they appear on the same host.
+	
 	- parameter oauth2: The OAuth2 instance to use for authorization when loading data.
+	- parameter host:   If given will handle redirects within the same host by way of `OAuth2DataLoaderSessionTaskDelegate`
 	*/
-	public init(oauth2: OAuth2) {
+	public init(oauth2: OAuth2, host: String? = nil) {
 		self.oauth2 = oauth2
 		super.init(logger: oauth2.logger)
+		if let host = host {
+			sessionDelegate = OAuth2DataLoaderSessionTaskDelegate(loader: self, host: host)
+		}
 	}
 	
 	
@@ -134,7 +140,7 @@ open class OAuth2DataLoader: OAuth2Requestable {
 					callback(response)
 				}
 			}
-				
+			
 			// some other error, pass along
 			catch {
 				callback(response)